requirements.txt
pandas==2.2.2
pillow==10.4.0
streamlit==1.38.0
